+++ apps/improve/src/app/modules/media-player/public-api.ts
@@ -1,0 +1,6 @@
+export { MediaPlayerModule } from './media-player.module';
+
+export {
+  AudioPlayerComponent,
+  IAudio,
+} from './components/audio-player/audio-player.component';