+++ apps/improve/src/app/modules/sounds/pages/list/list.component.ts
@@ -1,5 +1,11 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  OnInit,
+} from '@angular/core';
 import { SoundsService, ISound } from '../../../database/public-api';
+import { AudioPlayerComponent, IAudio } from '../../../media-player/public-api';
 
 @Component({
   selector: 'improve-themes-list',
@@ -12,10 +18,38 @@ export class ListComponent implements OnInit {
    * @internal
    */
   public list: ISound[] = [];
+  /**
+   * @internal
+   */
+  public sources: IAudio[] = [];
 
-  constructor(private sounds: SoundsService) {}
+  constructor(private sounds: SoundsService, private cdr: ChangeDetectorRef) {}
 
+  /**
+   * @internal
+   */
   public ngOnInit(): void {
     this.list = this.sounds.list();
+  }
+
+  /**
+   * @internal
+   */
+  public onPlay(item: ISound): void {
+    this.sources = [
+      {
+        url: item.url,
+        type: 'audio/mpeg',
+      },
+    ];
+
+    this.cdr.detectChanges();
+  }
+
+  /**
+   * @internal
+   */
+  public onSourceChange(component: AudioPlayerComponent): void {
+    component.play();
   }
 }