+++ apps/improve/src/app/modules/sounds/sounds.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SharedModule } from '../shared/shared.module';
+import { MediaPlayerModule } from '../media-player/media-player.module';
 import { ListComponent } from './pages/list/list.component';
 import { ItemComponent } from './components/item/item.component';
 import { SoundsRoutingModule } from './sounds-routing.module';
@@ -8,6 +9,6 @@ import { HomeComponent } from './pages/home/home.component';
 
 @NgModule({
   declarations: [ListComponent, ItemComponent, HomeComponent],
-  imports: [CommonModule, SharedModule, SoundsRoutingModule],
+  imports: [CommonModule, SharedModule, SoundsRoutingModule, MediaPlayerModule],
 })
 export class SoundsModule {}